feat(#81): split erros file
