+++ packages/validator/src/index.ts
@@ -1,4 +1,4 @@
-export const VALIDATOR_VERSION = "0.1.0";
-export * from "./lib/types";
-export * from "./lib/utils";
-export * from "./lib/errors";
+export const VALIDATOR_VERSION = '0.1.0';
+export * from './lib/types';
+export * from './lib/utils';
+export * from './lib/errors';
