+++ packages/validator/src/lib/errors/codes.ts
@@ -1,0 +1,69 @@
+export const ERROR_CODES = {
+  DATABASE_EMPTY_SCHEMA: 'DATABASE_EMPTY_SCHEMA',
+  SCHEMA_NOT_EXIST: 'SCHEMA_NOT_EXIST',
+  SCHEMA_INVALID: 'SCHEMA_INVALID',
+  SCHEMA_NAME_NOT_UNIQUE: 'SCHEMA_NAME_NOT_UNIQUE',
+  TABLE_NAME_NOT_INVALID: 'TABLE_NAME_NOT_INVALID',
+  TABLE_NAME_NOT_UNIQUE: 'TABLE_NAME_NOT_UNIQUE',
+  TABLE_NAME_NOT_CHANGED: 'TABLE_NAME_NOT_CHANGED',
+  TABLE_EMPTY_COLUMN: 'TABLE_EMPTY_COLUMN',
+  COLUMN_NOT_EXIST: 'COLUMN_NOT_EXIST',
+  COLUMN_NAME_NOT_UNIQUE: 'COLUMN_NAME_NOT_UNIQUE',
+  COLUMN_DATA_TYPE_REQUIRED: 'COLUMN_DATA_TYPE_REQUIRED',
+  COLUMN_DATA_TYPE_INVALID: 'COLUMN_DATA_TYPE_INVALID',
+  COLUMN_LENGTH_REQUIRED: 'COLUMN_LENGTH_REQUIRED',
+  COLUMN_PRECISION_REQUIRED: 'COLUMN_PRECISION_REQUIRED',
+  COLUMN_INVALID: 'COLUMN_INVALID',
+  COLUMN_INVALID_FORMAT: 'COLUMN_INVALID_FORMAT',
+  COLUMN_NAME_IS_RESERVED: 'COLUMN_NAME_IS_RESERVED',
+  MULTIPLE_AUTO_INCREMENT: 'MULTIPLE_AUTO_INCREMENT',
+  AUTO_INCREMENT_INVALID: 'AUTO_INCREMENT_INVALID',
+  DEFAULT_VALUE_INCOMPATIBLE: 'DEFAULT_VALUE_INCOMPATIBLE',
+  INDEX_PARSE_INVALID: 'INDEX_PARSE_INVALID',
+  INDEX_NAME_NOT_UNIQUE: 'INDEX_NAME_NOT_UNIQUE',
+  INDEX_TYPE_INVALID: 'INDEX_TYPE_INVALID',
+  INDEX_COLUMN_NOT_UNIQUE: 'INDEX_COLUMN_NOT_UNIQUE',
+  DUPLICATE_INDEX_DEFINITION: 'DUPLICATE_INDEX_DEFINITION',
+  INDEX_COLUMN_NOT_EXIST: 'INDEX_COLUMN_NOT_EXIST',
+  INDEX_COLUMN_SEQ_ERROR: 'INDEX_COLUMN_SEQ_ERROR',
+  INDEX_COLUMN_SORT_DIR_INVALID: 'INDEX_COLUMN_SORT_DIR_INVALID',
+  CONSTRAINT_NAME_NOT_UNIQUE: 'CONSTRAINT_NAME_NOT_UNIQUE',
+  CONSTRAINT_NULLABLE_CHANGE_NOT_ALLOWED: 'CONSTRAINT_NULLABLE_CHANGE_NOT_ALLOWED',
+  CONSTRAINT_COLUMN_REQUIRED: 'CONSTRAINT_COLUMN_REQUIRED',
+  CONSTRAINT_COLUMN_NOT_EXIST: 'CONSTRAINT_COLUMN_NOT_EXIST',
+  CONSTRAINT_COLUMN_NOT_UNIQUE: 'CONSTRAINT_COLUMN_NOT_UNIQUE',
+  PRIMARY_KEY_REQUIRED: 'PRIMARY_KEY_REQUIRED',
+  PRIMARY_KEY_NULLABLE: 'PRIMARY_KEY_NULLABLE',
+  DUPLICATE_KEY_DEFINITION: 'DUPLICATE_KEY_DEFINITION',
+  CHECK_EXPRESSION_REQUIRED: 'CHECK_EXPRESSION_REQUIRED',
+  DEFAULT_CONSTRAINT_ERROR: 'DEFAULT_CONSTRAINT_ERROR',
+  CONSTRAINT_COLUMN_SEQ_ERROR: 'CONSTRAINT_COLUMN_SEQ_ERROR',
+  RELATIONSHIP_COLUMN_SEQ_ERROR: 'RELATIONSHIP_COLUMN_SEQ_ERROR',
+  COLUMN_POSITION_SEQ_ERROR: 'COLUMN_POSITION_SEQ_ERROR',
+  IDENTIFYING_RELATIONSHIP_ORDER_ERROR: 'IDENTIFYING_RELATIONSHIP_ORDER_ERROR',
+  COMPOSITE_KEY_ORDER_MISMATCH: 'COMPOSITE_KEY_ORDER_MISMATCH',
+  RELATIONSHIP_NOT_EXIST: 'RELATIONSHIP_NOT_EXIST',
+  RELATIONSHIP_EMPTY_ERROR: 'RELATIONSHIP_EMPTY_ERROR',
+  RELATIONSHIP_NAME_NOT_UNIQUE: 'RELATIONSHIP_NAME_NOT_UNIQUE',
+  RELATIONSHIP_COLUNN_NOT_EXIST: 'RELATIONSHIP_COLUNN_NOT_EXIST',
+  RELATIONSHIP_COLUMN_MAPPING_DUPLICATE: 'RELATIONSHIP_COLUMN_MAPPING_DUPLICATE',
+  RELATIONSHIP_COLUMN_TYPE_INCOMPATIBLE: 'RELATIONSHIP_COLUMN_TYPE_INCOMPATIBLE',
+  RELATIONSHIP_DELETE_SET_NULL_ERROR: 'RELATIONSHIP_DELETE_SET_NULL_ERROR',
+  RELATIONSHIP_CYCLIC_REFERENCE: 'RELATIONSHIP_CYCLIC_REFERENCE',
+  RELATIONSHIP_TARGET_TABLE_NOT_EXIST: 'RELATIONSHIP_TARGET_TABLE_NOT_EXIST',
+  SCHEMA_IN_USE: 'SCHEMA_IN_USE',
+  TABLE_NOT_EXIST: 'TABLE_NOT_EXIST',
+  TABLE_IN_USE: 'TABLE_IN_USE',
+  COLUMN_IN_USE: 'COLUMN_IN_USE',
+  COLUMN_IN_PRIMARY_KEY: 'COLUMN_IN_PRIMARY_KEY',
+  INDEX_NOT_EXIST: 'INDEX_NOT_EXIST',
+  CONSTRAINT_NOT_EXIST: 'CONSTRAINT_NOT_EXIST',
+  CONSTRAINT_IN_USE: 'CONSTRAINT_IN_USE',
+  LAST_SCHEMA_CANNOT_DELETE: 'LAST_SCHEMA_CANNOT_DELETE',
+  LAST_COLUMN_CANNOT_DELETE: 'LAST_COLUMN_CANNOT_DELETE',
+  PRIMARY_KEY_CANNOT_DELETE: 'PRIMARY_KEY_CANNOT_DELETE',
+  COLUMN_TYPE_INCOMPATIBLE: 'COLUMN_TYPE_INCOMPATIBLE',
+  UNIQUE_SAME_AS_PRIMARY_KEY: 'UNIQUE_SAME_AS_PRIMARY_KEY',
+} as const;
+
+export type ErrorCode = (typeof ERROR_CODES)[keyof typeof ERROR_CODES];
