+++ packages/validator/src/lib/handlers/schemas.ts
@@ -3,20 +3,13 @@ import {
   SchemaNameInvalidError,
   SchemaNameNotUniqueError,
   DatabaseEmptySchemaError,
-} from "../errors";
-import { Database, SCHEMA, Schema } from "../types";
+} from '../errors';
+import { Database, SCHEMA, Schema } from '../types';
 
 export interface SchemaHandlers {
-  changeSchemaName: (
-    database: Database,
-    schemaId: Schema["id"],
-    newName: Schema["name"],
-  ) => Database;
-  createSchema: (
-    database: Database,
-    schema: Omit<Schema, "createdAt" | "updatedAt">,
-  ) => Database;
-  deleteSchema: (database: Database, schemaId: Schema["id"]) => Database;
+  changeSchemaName: (database: Database, schemaId: Schema['id'], newName: Schema['name']) => Database;
+  createSchema: (database: Database, schema: Omit<Schema, 'createdAt' | 'updatedAt'>) => Database;
+  deleteSchema: (database: Database, schemaId: Schema['id']) => Database;
 }
 
 export const schemaHandlers: SchemaHandlers = {
@@ -27,19 +20,14 @@ export const schemaHandlers: SchemaHandlers = {
     const result = SCHEMA.shape.name.safeParse(newName);
     if (!result.success) throw new SchemaNameInvalidError(newName);
 
-    const existingSchema = database.schemas.find(
-      (schema) => schema.name === newName && schema.id !== schemaId,
-    );
-    if (existingSchema)
-      throw new SchemaNameNotUniqueError(newName, existingSchema.id);
+    const existingSchema = database.schemas.find((schema) => schema.name === newName && schema.id !== schemaId);
+    if (existingSchema) throw new SchemaNameNotUniqueError(newName, existingSchema.id);
 
     return {
       ...database,
       isAffected: true,
       schemas: database.schemas.map((schema) =>
-        schema.id === schemaId
-          ? { ...schema, name: newName, isAffected: true }
-          : schema,
+        schema.id === schemaId ? { ...schema, name: newName, isAffected: true } : schema
       ),
     };
   },
@@ -48,8 +36,7 @@ export const schemaHandlers: SchemaHandlers = {
     if (!result.success) throw new SchemaNameInvalidError(schema.name);
 
     const existingSchema = database.schemas.find((s) => s.name === schema.name);
-    if (existingSchema)
-      throw new SchemaNameNotUniqueError(schema.name, existingSchema.id);
+    if (existingSchema) throw new SchemaNameNotUniqueError(schema.name, existingSchema.id);
 
     return {
       ...database,
